chat-service/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-filter==25.2
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
Markdown==3.10.2
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.12.1
sqlparse==0.5.5
